moved xml functions to separate xml module
